+++ Dart Challange/unit_test_3/bin/unit_test_3.dart
@@ -1,0 +1,5 @@
+import 'package:unit_test_3/check_prime_number.dart' as unit_test_3;
+
+void main(List<String> arguments) {
+  unit_test_3.checkPrimeNumberFunc(number: 2);
+}
